pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "dicom2img"
version="1.0.0"
dependencies = [
    "jsonargparse", "rich",
    "numpy==1.26", "matplotlib",
    "pydicom", "pillow"
] # project  dependencies

[project.scripts]
dicom="dicom2img.dicom:dicom_cli"
dicomdir="dicom2img.dicomdir:dicomdir_cli"

[tool.setuptools]
packages = ["dicom2img"]